requirements.txt
colorlog==6.7.0
httplib2==0.22.0
numpy==1.22.3
opencv_python==4.7.0.72
pafy==0.5.5
pydantic==1.10.2
python-dotenv==1.0.0
requests==2.27.1
numba==0.57.1